requirements.txt
setuptools~=75.1.0
requests~=2.32.3
Sphinx~=8.0.2
docutils~=0.21.2
pip~=24.2
numpy~=2.1.1
scikit-learn~=1.5.2
scipy~=1.14.1
six~=1.16.0
lxml~=5.3.0
matplotlib~=3.9.2
pillow~=10.4.0
protobuf~=5.28.1
fonttools~=4.54.0
pyparsing~=3.2.0b1
tornado~=6.4.1
GitPython~=3.1.43
omim-mwm~=10.3.0rc2
omim-data-essential~=10.3.0rc2
omim-data-files~=10.3.0rc2
omim-pygen~=10.3.0rc2
mwm~=0.10.1
beautifulsoup4~=4.13.0b2
requests-file~=2.1.0
filelock~=3.16.0
Wikipedia-API~=0.7.1
htmlmin~=0.1.12
Wikidata~=0.8.1
webdav~=1.1.7